+++ src/managers/config/defaults.ts
@@ -2,33 +2,32 @@ export default {
   config: {
     port: 3000,
     editor: {
-      mode: null, //[standalone, preview]
-      agent: null, //[vscode, theia]
+      mode: null, // [standalone, preview]
+      agent: null, // [vscode, theia]
       version: null,
     },
-    dirPath: "./.learn",
-    configPath: "./learn.json",
-    outputPath: "./.learn/dist",
-    publicPath: "/preview",
+    dirPath: './.learn',
+    configPath: './learn.json',
+    outputPath: './.learn/dist',
+    publicPath: '/preview',
     publicUrl: null,
-    exerciseNameValidationRegex: "/^(d{2,3}(.d{1,2})?-([dA-Za-z]+(-|_)?)+)$/",
-    language: "auto",
-    grading: "isolated", // [isolated, incremental]
-    exercisesPath: "./", // path to the folder that contains the exercises
+    language: 'auto',
+    grading: 'isolated', // [isolated, incremental]
+    exercisesPath: './', // path to the folder that contains the exercises
     webpackTemplate: null, // if you want webpack to use an HTML template
     disableGrading: false,
-    disabledActions: [], //Possible: 'build', 'test' or 'reset'
+    disabledActions: [], // Possible: 'build', 'test' or 'reset'
     actions: [], // ⚠️ deprecated, leave empty )
     entries: {
-      html: "index.html",
-      vanillajs: "index.js",
-      react: "app.jsx",
-      node: "app.js",
-      python3: "app.py",
-      java: "app.java",
+      html: 'index.html',
+      vanillajs: 'index.js',
+      react: 'app.jsx',
+      node: 'app.js',
+      python3: 'app.py',
+      java: 'app.java',
     },
   },
-  address: "http://localhost",
+  address: 'http://localhost',
   currentExercise: null,
   exercises: [],
-};
+}